// core/admin-dashboard/store/admin/action-types.ts
export enum ActionTypes {
  GET_ADMINS = "GET_ADMINS",
  GET_ADMIN = "GET_ADMIN",
  CREATE_ADMIN = "CREATE_ADMIN",
  UPDATE_ADMIN = "UPDATE_ADMIN",
  DELETE_ADMIN = "DELETE_ADMIN",
  HARD_DELETE_ADMIN = "HARD_DELETE_ADMIN",
  ENABLE_AUTO_CENSORSHIP_POST = "ENABLE_AUTO_CENSORSHIP_POST",
  RESTORE_ADMIN = "RESTORE_ADMIN",
  DISABLE_AUTO_CENSORSHIP_POST = "DISABLE_AUTO_CENSORSHIP_POST",
  UPDATE_ADMIN_PASSWORD = "UPDATE_ADMIN_PASSWORD",
  UPDATE_ADMIN_PERSONAL_PASSWORD = "UPDATE_ADMIN_PERSONAL_PASSWORD",
  GET_ADMIN_ANALYTICS = "GET_ADMIN_ANALYTICS",
  RESET_ADMIN_LOGIN_FAILED_TIMES = "RESET_ADMIN_LOGIN_FAILED_TIMES",
  BATCH_UPLOAD_ADMIN = "BATCH_UPLOAD_ADMIN",
}